code/python/requirements.txt
# --- Core runtime (Cloud Run) ---
aiohttp>=3.9.1
python-dotenv>=1.0.0
httpx>=0.28.1
openai>=1.12.0
pyyaml>=6.0.1
feedparser>=6.0.1
chardet>=5.2.0              # needed by core/utils/trim_schema_json.py
jsonschema>=4.19.1
marshmallow<4.0.0
packaging>=23,<24
protobuf>=3.20,<5
mcp>=1.0.0

# --- Optional: Azure Search integration (keep if you use NLWeb's Azure retrievers) ---
azure-identity>=1.15.0
azure-search-documents>=11.4.0

# --- LLM providers (you already pre-install OpenAI above) ---
# If you plan to use others, uncomment to preinstall and avoid runtime pip in Cloud Run.
# anthropic>=0.18.1
# google-genai>=0.7.1
# google-cloud-aiplatform>=1.38.0
# huggingface_hub>=0.31.0
# azure-ai-inference>=1.0.0b9
# azure-core>=1.30.0

# --- Retrieval backends (uncomment only if you actually use them) ---
# qdrant-client>=1.14.0
# pymilvus>=1.1.0
# numpy
# elasticsearch[async]>=8,<9
# httpx>=0.28.1  # already listed above
# ollama>=0.5.1
# psycopg[binary]>=3.1.12
# psycopg[pool]>=3.2.0
# pgvector>=0.4.0

# --- Dev/analysis only (move to requirements-dev.txt instead of installing in prod) ---
# matplotlib>=3.8.0
# seaborn>=0.13.0
